Urls And Views CleanUp and Management
