+++ babble_django/urls.py
@@ -20,9 +20,9 @@ from django.urls import path, re_path, include
 from django.views.generic import TemplateView
 
 from babbles.views import (
-    local_tweets_list_view,
-    local_tweets_detail_view,
-    local_tweets_profile_view,
+    tweets_list_view,
+    tweets_detail_view,
+    tweets_profile_view,
 )
 
 urlpatterns = [
@@ -33,10 +33,10 @@ urlpatterns = [
     # path('babbles/',tweet_list_view),
     # path('create-tweet',tweet_create_view),
 
-    path('', local_tweets_list_view),
-    path('<int:tweet_id>', local_tweets_detail_view),
-    path('profile/<str:username>', local_tweets_profile_view),
-    path('api/babbles/',include('babbles.urls'))
+    path('', tweets_list_view),
+    path('<int:tweet_id>', tweets_detail_view),
+    path('profile/<str:username>', tweets_profile_view),
+    path('api/babbles/',include('babbles.api.urls'))
 ]
 
 if settings.DEBUG:
