+++ babbles/api/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path, re_path
 
 from .views import (
-    home_view,
+
     tweet_detail_view,
     tweet_list_view, 
     tweet_create_view,
